pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-wsvnc"
version = "1.0.0"
description = "WebSocket VNC Client."
authors = [{ name = "Matthew Boyer", email = "matthew.boyer@cynnovative.com" }]
dependencies = [
  "websockets~=12.0",
  "pillow~=10.4.0",
  "pycryptodomex~=3.19.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/Cynnovative/py-wsvnc"

[project.optional-dependencies]
dev = ["docformatter[tomli]>=1.7.5", "mypy", "pytest>=7.4.1", "ruff>=0.1.6", "websockify>=0.11.0", "types-Pillow>=10.1.0.1", "coverage>=7.3.2"]

[tool.isort]
profile = "black"

[tool.ruff.isort]
known-first-party = ["wsvnc"]

[tool.mypy]
ignore_missing_imports = false
disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
files = "src/"
exclude = "tests/"
follow_imports = "silent"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/mock_server_tests"]
log_cli = true
log_cli_level = "INFO"

[tool.docformatter]
wrap-summaries = 88
wrap-descriptions = 88
black = true
exclude = ["notebooks", "_version.py"]
recursive = true

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff]
extend-select = ["D", "I"]
extend-exclude = [
  "_version.py",
  "notebooks",
]
ignore = [
  "D101",
  "D102",
]

[tool.ruff.per-file-ignores]
"tests/**/test_*.py" = ["D"]
